requirements.txt
amqp==5.1.0
asgiref==3.4.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
bzip2==1.0.8
ca-certificates==2022.2.1
celery==5.2.3
certifi==2021.5.30
cffi==1.15.0
charset-normalizer==2.0.12
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
click==8.0.4
cryptography==36.0.0
django-bootstrap4==22.1
django-celery-beat==2.2.1
django-celery-results==2.3.0
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-filter==21.1
django-simpleui==2022.3.19
django-timezone-field==4.2.3
django==3.2.5
djangorestframework==3.13.1
idna==3.3
kombu==5.2.4
krb5==1.19.2
libcxx==12.0.0
libedit==3.1.20210910
libffi==3.3
libpq==12.9
mysql-connector-c==6.1.11
mysqlclient==2.0.3
naked==0.1.31
ncurses==6.3
openssl==1.1.1m
pip==21.2.4
prompt-toolkit==3.0.28
psycopg2==2.8.6
pycparser==2.21
pycryptodome==3.14.1
pymysql==1.0.2
python-crontab==2.6.0
python-dateutil==2.8.2
python==3.10.0
pytz==2021.3
pyyaml==6.0
readline==8.1.2
requests==2.27.1
setuptools==59.6.0
shellescape==3.8.1
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.1
sqlalchemy==1.4.27
sqlite==3.37.2
sqlparse==0.4.1
tk==8.6.11
typing_extensions==3.10.0.2
tzdata==2021e
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wheel==0.37.1
xz==5.2.5
zlib==1.2.11
